djangoAI-main/requirements.txt
langchain
torch
accelerate
# ctransformers
huggingface_hub
langchain_community
langchain_huggingface
langchain-google-genai
langchain_core
openai
python-dotenv
Django
# streamlit
diffusers
# !pip install git+https://github.com/suno-ai/bark.git
langdetect
langcodes
#for chromadb
chromadb
langchain_chroma
faiss_cpu
unstructured
python-docx
tika
sentence_transformers
pypdf
accelerate


langchain-google-genai
scikit-learn
bark